extract artwork class into seperate file
